+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Redirect endpoint to receive TikTok's redirect
+- Middleware for parsing AuthCode and state params
 
 ### Changed